backend changes

+++ backend/app/schemas/dashboard.py
@@ -37,7 +37,7 @@ class AlertDetail(BaseModel):
     dst_ip: Optional[str] = None
     protocol: Optional[str] = None
     byte_count: Optional[int] = None
-    packet_size: Optional[int] = None
+    packet_size: Optional[float] = None
     # Joined from event_context
     src_mac: Optional[str] = None
     dst_mac: Optional[str] = None
